+++ backend/canisters/community/impl/src/jobs/import_groups.rs
@@ -107,7 +107,7 @@ fn deserialize_group(group_id: ChatId) {
 // For users who are not members, lookup their principals, then join them to the community, then add
 // them to the default channels, then add the new channel to their set of channels.
 async fn process_channel_members(group_id: ChatId, channel_id: ChannelId, attempt: u32) {
-    let (members_to_add, local_user_index_canister_id, is_public) = mutate_state(|state| {
+    let (members_to_add, local_user_index_canister_id) = mutate_state(|state| {
         let channel = state.data.channels.get(&channel_id).unwrap();
         let mut to_add = Vec::new();
         for user_id in channel.chat.members.iter().map(|m| m.user_id) {
@@ -118,7 +118,7 @@ async fn process_channel_members(group_id: ChatId, channel_id: ChannelId, attemp
             }
         }
 
-        (to_add, state.data.local_user_index_canister_id, state.data.is_public)
+        (to_add, state.data.local_user_index_canister_id)
     });
 
     let mut members_added = Vec::new();
@@ -132,7 +132,6 @@ async fn process_channel_members(group_id: ChatId, channel_id: ChannelId, attemp
         {
             mutate_state(|state| {
                 let now = state.env.now();
-                let notifications_muted = is_public;
                 let default_channel_ids = state.data.channels.default_channel_ids();
 
                 for (user_id, principal) in users {
@@ -143,7 +142,9 @@ async fn process_channel_members(group_id: ChatId, channel_id: ChannelId, attemp
                             let member = state.data.members.get_by_user_id_mut(&user_id).unwrap();
                             for default_channel_id in default_channel_ids.iter() {
                                 if let Some(channel) = state.data.channels.get_mut(default_channel_id) {
-                                    join_channel_unchecked(channel, member, notifications_muted, now);
+                                    if channel.chat.gate.is_none() {
+                                        join_channel_unchecked(channel, member, true, true, now);
+                                    }
                                 }
                             }
                             member.channels.insert(channel_id);